fix: rename `/ResourceList` folder in `/useResourceList`
