+++ packages/app-elements/src/ui/forms/InputResourceGroup/FullList.tsx
@@ -5,7 +5,7 @@ import { Spacer } from '#ui/atoms/Spacer'
 import { Text } from '#ui/atoms/Text'
 import { SearchBar } from '#ui/composite/SearchBar'
 import { type OverlayProps } from '#ui/internals/Overlay'
-import { useResourceList } from '#ui/resources/ResourceList'
+import { useResourceList } from '#ui/resources/useResourceList'
 import { type ListableResourceType, type QueryFilter } from '@commercelayer/sdk'
 import isEmpty from 'lodash/isEmpty'
 import { useCallback, useEffect, useMemo, useState } from 'react'
